+++ mediation/UnityAds/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unity Ads Unity Mediation Plugin Changelog
 
+#### Version Next
+- Renamed `DummyClient` to `PlaceholderClient`.
+
 #### [Version 3.12.1](https://dl.google.com/googleadmobadssdk/mediation/unity/unity/UnityAdsUnityAdapter-3.12.1.zip)
 - Supports [Unity Ads Android adapter version 4.12.1.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/main/ThirdPartyAdapters/unity/CHANGELOG.md#version-41210).
 - Supports [Unity Ads iOS adapter version 4.12.1.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/main/adapters/Unity/CHANGELOG.md#version-41210).